filepath to datapath
