+++ tests/transform/boundary_conditions/test_cams.py
@@ -2,7 +2,6 @@ from helpers import get_bc_datapath
 import logging
 import numpy as np
 
-from openghg.transform import transform_bc_data
 from openghg.transform.boundary_conditions import parse_cams
 
 mpl_logger = logging.getLogger("matplotlib")
@@ -60,30 +59,3 @@ def test_parse_cams():
 
     assert bc_data.time.size == 3
     assert bc_data.time.values[0] == np.datetime64("2022-01-01T00:00")
-
-
-def test_transform_cams_n2o_bc():
-    "Test CAMS parser for transform_boundary_conditions"
-    bc_input = "cams_test"
-    cams_version = "v22r1"
-    domain = "europe"
-    species = "n2o"
-    period = "daily"
-    filename = "cams73_v22r1_n2o_test_202201.nc"
-    data_path = get_bc_datapath(filename=filename)
-
-    results = transform_bc_data(
-        data_path=data_path,
-        species=species,
-        bc_input=bc_input,
-        period=period,
-        cams_version=cams_version,
-        domain=domain,
-        source_format="cams",
-        store="user",
-    )
-
-    expected_metadata = {"species": species, "domain": domain, "bc_input": bc_input, "file": filename}
-
-    for k, v in expected_metadata.items():
-        assert results[0][k].lower() == v.lower()
